Switch to toggle button for the elevator.
