--- src/main/java/frc/team4069/robot/commands/ElevatorStartCommand.java
@@ -1,14 +1,0 @@
-package frc.team4069.robot.commands;
-
-public class ElevatorStartCommand extends CommandBase {
-
-    @Override
-    protected void initialize() {
-        elevator.start();
-    }
-
-    @Override
-    protected boolean isFinished() {
-        return true;
-    }
-}
