--- src/main/java/frc/team4069/robot/commands/ElevatorStopCommand.java
@@ -1,14 +1,0 @@
-package frc.team4069.robot.commands;
-
-public class ElevatorStopCommand extends CommandBase {
-
-    @Override
-    protected void initialize() {
-        elevator.stop();
-    }
-
-    @Override
-    protected boolean isFinished() {
-        return true;
-    }
-}
