+++ src/main/java/frc/team4069/robot/commands/ElevatorToggleCommand.java
@@ -1,0 +1,18 @@
+package frc.team4069.robot.commands;
+
+public class ElevatorToggleCommand extends CommandBase {
+
+    @Override
+    protected void initialize() {
+        if(elevator.isStarted()) {
+            elevator.stop();
+        }else {
+            elevator.start();
+        }
+    }
+
+    @Override
+    protected boolean isFinished() {
+        return true;
+    }
+}
